+++ inbox/constants.py
@@ -28,34 +28,32 @@ class MessageLogStatus(enum.Enum):
     NEW = 1
     QUEUED = 2
     SENT = 3
-    SKIPPED_FOR_PREF = 4
+    NOT_SENDABLE = 4
     FAILED = 5
 
     __labels__ = {
         NEW: 'New',
         QUEUED: 'Queued',
         SENT: 'Sent',
-        SKIPPED_FOR_PREF: 'Skipped for preference',
-        FAILED: 'Failed'
+        NOT_SENDABLE: 'Not Sendable',
+        FAILED: 'Failed',
     }
 
 
-class MessageLogFailureReason(enum.Enum):
+class MessageLogStatusReason(enum.Enum):
 
     MISSING_TEMPLATE = 101
-    MISSING_APP_PUSH_KEY = 102
+    MISSING_ID = 102
 
-    EMAIL_NOT_VERIFIED = 201
-    SMS_NOT_VERIFIED = 202
+    NOT_VERIFIED = 201
 
-    UNKNOWN_EXCEPTION = 301
+    PREFERENCE_OFF = 301
 
     __labels__ = {
         MISSING_TEMPLATE: 'Missing template',
-        MISSING_APP_PUSH_KEY: 'Missing App Push Key',
+        MISSING_ID: 'Missing id',
 
-        EMAIL_NOT_VERIFIED: 'Email is not verified',
-        SMS_NOT_VERIFIED: 'SMS is not verified',
+        NOT_VERIFIED: 'Not verified',
 
-        UNKNOWN_EXCEPTION: 'Unknown Exception'
+        PREFERENCE_OFF: 'Pref off'
     }