+++ tests/hooks/default.py
@@ -1,5 +1,5 @@
 from inbox import settings as inbox_settings
-from inbox.constants import MessageLogStatus, MessageMedium, MessageLogFailureReason
+from inbox.constants import MessageLogStatus, MessageMedium, MessageLogStatusReason
 from inbox.models import MessageLog
 
 
@@ -9,20 +9,20 @@ def can_send(message_log: MessageLog):
 
     if message_log.medium == MessageMedium.APP_PUSH:
         if not user.notification_key:
-            message_log.status = MessageLogStatus.FAILED
-            message_log.failure_reason = MessageLogFailureReason.MISSING_APP_PUSH_KEY
+            message_log.status = MessageLogStatus.NOT_SENDABLE
+            message_log.status_reason = MessageLogStatusReason.MISSING_ID
             return False
 
     if message_log.medium == MessageMedium.EMAIL:
         if inbox_settings.get_config()['CHECK_IS_EMAIL_VERIFIED'] and not user.is_email_verified:
-            message_log.status = MessageLogStatus.FAILED
-            message_log.failure_reason = MessageLogFailureReason.EMAIL_NOT_VERIFIED
+            message_log.status = MessageLogStatus.NOT_SENDABLE
+            message_log.status_reason = MessageLogStatusReason.NOT_VERIFIED
             return False
 
     if message_log.medium == MessageMedium.SMS:
         if inbox_settings.get_config()['CHECK_IS_SMS_VERIFIED'] and not user.is_sms_verified:
-            message_log.status = MessageLogStatus.FAILED
-            message_log.failure_reason = MessageLogFailureReason.SMS_NOT_VERIFIED
+            message_log.status = MessageLogStatus.NOT_SENDABLE
+            message_log.status_reason = MessageLogStatusReason.NOT_VERIFIED
             return False
 
     preference = next((g for g in user.message_preferences.groups if g['id'] == message_group['id']))
@@ -30,6 +30,7 @@ def can_send(message_log: MessageLog):
     if preference.get(message_log.medium.name.lower()):
         return True
     else:
-        message_log.status = MessageLogStatus.SKIPPED_FOR_PREF
+        message_log.status = MessageLogStatus.NOT_SENDABLE
+        message_log.status_reason = MessageLogStatusReason.PREFERENCE_OFF
 
     return False